Smart-Flower-Pot
The aim of the project is to determine the moisture needs of a plant by analyzing data received from a soil moisture sensor using an Arduino Nano and comparing it to a predefined threshold value. This information is then communicated to the user through visual and textual messages. The system monitors the soil moisture level and displays messages such as "Water me", "I don't need water", and "Too much water" on an LCD screen based on the moisture level. Additionally, red, green, and blue LEDs are used to convey the plant's status more quickly and visually. The red LED indicates the need for water, the green LED signifies that the plant is in good condition, and the blue LED indicates excess water. This allows users to easily understand the water needs of their plants and keep them healthy.